requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appnope==0.1.4
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-generator==1.10
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-spinner==0.1.10
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.0
croniter==1.4.1
cryptography==42.0.2
cycler==0.12.1
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
dnspython==2.5.0
ecdsa==0.18.0
effdet==0.4.1
email-validator==2.1.0.post1
emoji==2.10.1
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.109.2
filelock==3.13.1
filetype==1.2.0
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2024.2.0
gotrue==2.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
iopath==0.1.10
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
langchain==0.1.11
langchain-community==0.0.25
langchain-core==0.1.29
langchain-experimental==0.0.50
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langchainhub==0.1.14
langdetect==1.0.9
langgraph==0.0.26
langsmith==0.1.19
layoutparser==0.3.4
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.12.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
orjson==3.9.15
packaging==23.2
pandas==2.2.0
parso==0.8.3
pdf2docx==0.5.8
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pexpect==4.9.0
pikepdf==8.12.0
pillow==10.2.0
pillow_heif==0.15.0
platformdirs==4.2.0
portalocker==2.8.2
postgrest==0.16.1
prisma==0.12.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pypandoc==1.12
pyparsing==3.1.1
pypdf==4.0.1
pypdfium2==4.27.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.6.1
realtime==1.0.2
redbird==0.7.1
redis==5.0.2
regex==2023.12.25
requests==2.31.0
rsa==4.9
safetensors==0.4.2
scipy==1.12.0
simsimd==3.9.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.26
stack-data==0.6.3
starlette==0.36.3
storage3==0.7.0
StrEnum==0.4.15
supabase==2.4.0
supafunc==0.3.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.15.2
tomlkit==0.12.3
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.37.2
types-requests==2.31.0.20240125
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
unstructured==0.12.4
unstructured-client==0.18.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
upstash-redis==1.0.0
urllib3==2.2.0
uvicorn==0.27.1
wcwidth==0.2.13
websockets==11.0.3
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.4
